refactor: update file hierarchy for Community directory
